Move functions from AddCommandParser into ParserUtil
