+++ src/main/java/seedu/address/logic/parser/AddCommandParser.java
@@ -1,7 +1,5 @@
 package seedu.address.logic.parser;
 
-import static seedu.address.logic.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
-
 import java.util.HashMap;
 
 import seedu.address.logic.commands.AddCommand;
@@ -23,40 +21,12 @@ public class AddCommandParser implements Parser<AddCommand> {
     public AddCommand parse(String args) throws ParseException {
         args = args.trim();
         String[] parts = args.split("/", -1);
-        parts = removeFirstItemFromStringList(parts);
-        requireValidParts(parts);
-        HashMap<String, String> attributeMap = getAttributeHashMapFromAttributeStrings(parts);
+        parts = ParserUtil.removeFirstItemFromStringList(parts);
+        HashMap<String, String> attributeMap = ParserUtil.getAttributeHashMapFromAttributeStrings(parts);
         return new AddCommand(attributeMap);
     }
 
-    private String[] removeFirstItemFromStringList(String[] parts) {
-        // Solution below generated by GitHub Copilot
-        String[] newParts;
-        newParts = new String[parts.length - 1];
-        System.arraycopy(parts, 1, newParts, 0, parts.length - 1);
-        return newParts;
-    }
 
-    private static HashMap<String, String> getAttributeHashMapFromAttributeStrings(String[] parts) {
-        HashMap<String, String> attributeMap = new HashMap<>();
-        for (int i = 0; i < parts.length; i++) {
-            String[] attribute = separateAttributeNamesAndValues(parts[i]);
-            String attributeName = attribute[0];
-            String attributeValue = attribute[1];
-            attributeMap.put(attributeName, attributeValue);
-        }
-        return attributeMap;
-    }
 
-    private void requireValidParts(String[] parts) throws ParseException {
-        for (int i = 0; i < parts.length; i++) {
-            if (separateAttributeNamesAndValues(parts[i]).length != 2) {
-                throw new ParseException(String.format(MESSAGE_INVALID_COMMAND_FORMAT, AddCommand.MESSAGE_USAGE));
-            }
-        }
-    }
 
-    private static String[] separateAttributeNamesAndValues(String parts) {
-        return parts.trim().split(" ", 2);
-    }
 }
