+++ src/main/java/seedu/address/logic/parser/ParserUtil.java
@@ -1,14 +1,17 @@
 package seedu.address.logic.parser;
 
 import static java.util.Objects.requireNonNull;
+import static seedu.address.logic.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 
 import java.util.Collection;
+import java.util.HashMap;
 import java.util.HashSet;
 import java.util.Set;
 
 import seedu.address.commons.core.index.Index;
 import seedu.address.commons.util.StringUtil;
 import seedu.address.logic.Messages;
+import seedu.address.logic.commands.AddCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.person.Address;
 import seedu.address.model.person.Email;
@@ -135,5 +138,48 @@ public class ParserUtil {
             throw new ParseException(Messages.MESSAGE_INVALID_PERSON_UUID);
         }
         return trimmedUuid;
+    }
+
+    /**
+     * Parses attributes in command arguments into a HashMap representing the pairs of attribute names and values
+     *
+     * @param parts The pairs of attribute names and values in the command arguments
+     * @return  A HashMap containing the pairs of attribute names and values
+     */
+    public static HashMap<String, String> getAttributeHashMapFromAttributeStrings(String[] parts)
+            throws ParseException {
+        requireValidParts(parts);
+        HashMap<String, String> attributeMap = new HashMap<>();
+        for (int i = 0; i < parts.length; i++) {
+            String[] attribute = separateAttributeNamesAndValues(parts[i]);
+            String attributeName = attribute[0];
+            String attributeValue = attribute[1];
+            attributeMap.put(attributeName, attributeValue);
+        }
+        return attributeMap;
+    }
+    /**
+     * Removes the first item from a string list
+     *
+     * @param parts The string list
+     * @return A new string list with the first item removed
+     */
+    public static String[] removeFirstItemFromStringList(String[] parts) {
+        // Solution below generated by GitHub Copilot
+        String[] newParts;
+        newParts = new String[parts.length - 1];
+        System.arraycopy(parts, 1, newParts, 0, parts.length - 1);
+        return newParts;
+    }
+
+    private static String[] separateAttributeNamesAndValues(String parts) {
+        return parts.trim().split(" ", 2);
+    }
+    private static void requireValidParts(String[] parts) throws ParseException {
+        for (int i = 0; i < parts.length; i++) {
+            if (separateAttributeNamesAndValues(parts[i]).length != 2) {
+                throw new ParseException(String.format(MESSAGE_INVALID_COMMAND_FORMAT, AddCommand.MESSAGE_USAGE));
+            }
+        }
     }
 }
